requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
scikit-image==0.22.0
gunicorn==21.2.0
